+++ tests/e2e/tx/service_test.go
@@ -30,6 +30,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 	authclient "github.com/cosmos/cosmos-sdk/x/auth/client"
 	authtest "github.com/cosmos/cosmos-sdk/x/auth/client/testutil"
+	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 )
@@ -772,6 +773,7 @@ func (s IntegrationTestSuite) TestGetBlockWithTxs_GRPCGateway() {
 }
 
 func (s IntegrationTestSuite) TestTxEncode_GRPC() {
+	val := s.network.Validators[0]
 	txBuilder := s.mkTxBuilder()
 	protoTx, err := txBuilderToProtoTx(txBuilder)
 	s.Require().NoError(err)
@@ -798,6 +800,10 @@ func (s IntegrationTestSuite) TestTxEncode_GRPC() {
 			} else {
 				s.Require().NoError(err)
 				s.Require().NotEmpty(res.GetTxBytes())
+
+				tx, err := val.ClientCtx.TxConfig.TxDecoder()(res.TxBytes)
+				s.Require().NoError(err)
+				s.Require().Equal(protoTx.GetMsgs(), tx.GetMsgs())
 			}
 		})
 	}
@@ -829,7 +835,13 @@ func (s *IntegrationTestSuite) TestTxEncode_GRPCGateway() {
 			if tc.expErr {
 				s.Require().Contains(string(res), tc.expErrMsg)
 			} else {
+				var result tx.TxEncodeResponse
+				err := val.ClientCtx.Codec.UnmarshalJSON(res, &result)
 				s.Require().NoError(err)
+
+				tx, err := val.ClientCtx.TxConfig.TxDecoder()(result.TxBytes)
+				s.Require().NoError(err)
+				s.Require().Equal(protoTx.GetMsgs(), tx.GetMsgs())
 			}
 		})
 	}
@@ -867,6 +879,11 @@ func (s IntegrationTestSuite) TestTxDecode_GRPC() {
 			} else {
 				s.Require().NoError(err)
 				s.Require().NotEmpty(res.GetTx())
+
+				txb := authtx.WrapTx(res.Tx)
+				tx, err := val.ClientCtx.TxConfig.TxEncoder()(txb.GetTx())
+				s.Require().NoError(err)
+				s.Require().Equal(encodedTx, tx)
 			}
 		})
 	}
@@ -876,10 +893,10 @@ func (s IntegrationTestSuite) TestTxDecode_GRPCGateway() {
 	val := s.network.Validators[0]
 	txBuilder := s.mkTxBuilder()
 
-	txBytes, err := val.ClientCtx.TxConfig.TxEncoder()(txBuilder.GetTx())
+	encodedTxBytes, err := val.ClientCtx.TxConfig.TxEncoder()(txBuilder.GetTx())
 	s.Require().NoError(err)
 
-	invalidTxBytes := append(txBytes, byte(0o00))
+	invalidTxBytes := append(encodedTxBytes, byte(0o00))
 
 	testCases := []struct {
 		name      string
@@ -889,7 +906,7 @@ func (s IntegrationTestSuite) TestTxDecode_GRPCGateway() {
 	}{
 		{"empty request", &tx.TxDecodeRequest{}, true, "invalid empty tx bytes"},
 		{"invalid tx bytes", &tx.TxDecodeRequest{TxBytes: invalidTxBytes}, true, "tx parse error"},
-		{"valid request with tx_bytes", &tx.TxDecodeRequest{TxBytes: txBytes}, false, ""},
+		{"valid request with tx_bytes", &tx.TxDecodeRequest{TxBytes: encodedTxBytes}, false, ""},
 	}
 
 	for _, tc := range testCases {
@@ -902,7 +919,14 @@ func (s IntegrationTestSuite) TestTxDecode_GRPCGateway() {
 			if tc.expErr {
 				s.Require().Contains(string(res), tc.expErrMsg)
 			} else {
+				var result tx.TxDecodeResponse
+				err := val.ClientCtx.Codec.UnmarshalJSON(res, &result)
 				s.Require().NoError(err)
+
+				txb := authtx.WrapTx(result.Tx)
+				tx, err := val.ClientCtx.TxConfig.TxEncoder()(txb.GetTx())
+				s.Require().NoError(err)
+				s.Require().Equal(encodedTxBytes, tx)
 			}
 		})
 	}
@@ -924,6 +948,7 @@ func (s IntegrationTestSuite) TestTxEncodeAmino_GRPC() {
 	}{
 		{"nil request", nil, true, "request cannot be nil"},
 		{"empty request", &tx.TxEncodeAminoRequest{}, true, "invalid empty tx json"},
+		{"invalid request", &tx.TxEncodeAminoRequest{AminoJson: "invalid tx json"}, true, "invalid request"},
 		{"valid request with amino-json", &tx.TxEncodeAminoRequest{AminoJson: string(txJSONBytes)}, false, ""},
 	}
 
@@ -938,6 +963,11 @@ func (s IntegrationTestSuite) TestTxEncodeAmino_GRPC() {
 			} else {
 				s.Require().NoError(err)
 				s.Require().NotEmpty(res.GetAminoBinary())
+
+				stdTxConfig := legacytx.StdTxConfig{Cdc: val.ClientCtx.LegacyAmino}
+				newStdTx, err := stdTxConfig.TxJSONDecoder()(res.GetAminoBinary())
+				s.Require().NoError(err)
+				s.Require().Equal(newStdTx.GetMsgs(), stdTx.GetMsgs())
 			}
 		})
 	}
@@ -958,6 +988,7 @@ func (s *IntegrationTestSuite) TestTxEncodeAmino_GRPCGateway() {
 		expErrMsg string
 	}{
 		{"empty request", &tx.TxEncodeAminoRequest{}, true, "invalid empty tx json"},
+		{"invalid request", &tx.TxEncodeAminoRequest{AminoJson: "invalid tx json"}, true, "invalid request"},
 		{"valid request with amino-json", &tx.TxEncodeAminoRequest{AminoJson: string(txJSONBytes)}, false, ""},
 	}
 
@@ -971,7 +1002,14 @@ func (s *IntegrationTestSuite) TestTxEncodeAmino_GRPCGateway() {
 			if tc.expErr {
 				s.Require().Contains(string(res), tc.expErrMsg)
 			} else {
+				var result tx.TxEncodeAminoResponse
+				err := val.ClientCtx.Codec.UnmarshalJSON(res, &result)
 				s.Require().NoError(err)
+
+				stdTxConfig := legacytx.StdTxConfig{Cdc: val.ClientCtx.LegacyAmino}
+				newStdTx, err := stdTxConfig.TxJSONDecoder()(result.GetAminoBinary())
+				s.Require().NoError(err)
+				s.Require().Equal(newStdTx.GetMsgs(), stdTx.GetMsgs())
 			}
 		})
 	}